requirements.txt
langchain
streamlit
streamlit-chat
python-dotenv
openai
faiss-cpu
langchain-community
tiktoken
PyPDF2
docling